+++ packages/iov-core/src/multichainsigner.ts
@@ -108,6 +108,13 @@ export class MultiChainSigner {
   }
 
   /**
+   * A chain-dependent validation of address
+   */
+  public isValidAddress(chainId: ChainId, address: string): boolean {
+    return this.getChain(chainId).codec.isValidAddress(address);
+  }
+
+  /**
    * Queries the nonce, signs the transaction and posts it to the blockchain.
    *
    * The transaction signer is determined by the transaction content. A lookup for