+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   in @iov/ethereum and @iov/tendermint-rpc later on.
 * @iov/keycontrol: Add `HdPaths.bip44Like` and `HdPaths.iov`
 * @iov/iov-core: Add `MultiChainSigner.shutdown` to shutdown the signer.
+* @iov/iov-core: Add `MultiChainSigner.isValidAddress` for chain-specific address input validation.
 * @iov/stream: Add an implementation of `concat` that buffers stream events
 
 Breaking changes